data plumbing between App and the reducers
